refatoracao search bar
