+++ src/components/Home.js
@@ -4,10 +4,27 @@ import CategoriesAside from './CategoriesAside';
 import SearchBar from './SearchBar';
 
 class Home extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      category: '',
+    };
+    this.handleChange = this.handleChange.bind(this);
+  }
+
+  handleChange({ target }) {
+    this.setState({
+      category: target.value,
+    });
+  }
+
   render() {
+    const { category } = this.state;
+
     return (
       <div>
-        <SearchBar />
+        <SearchBar category={ category } />
         <p data-testid="home-initial-message">
           Digite algum termo de pesquisa ou escolha uma categoria.
         </p>
@@ -20,7 +37,7 @@ class Home extends React.Component {
           </button>
         </Link>
         <aside>
-          <CategoriesAside />
+          <CategoriesAside value={ category } handleChange={ this.handleChange } />
         </aside>
       </div>
     );
